Rename jax-rs-* modules
