+++ instrumentation/dropwizard-testing/dropwizard-testing.gradle
@@ -10,7 +10,7 @@ apply from: "${rootDir}/gradle/java.gradle"
 
 dependencies {
   testCompile project(':instrumentation:java-concurrent')
-  testCompile project(':instrumentation:jax-rs-annotations-2.0')
+  testCompile project(':instrumentation:jaxrs-2.0')
   testCompile project(':instrumentation:servlet:request-3.0')
 
   // First version with DropwizardTestSupport:
