--- instrumentation/jax-rs-annotations-2.0/filter-jersey-2.0/filter-jersey-2.0.gradle
@@ -1,17 +1,0 @@
-muzzle {
-  // Cant assert fails because muzzle assumes all instrumentations will fail
-  // Instrumentations in jax-rs-annotations-2 will pass
-  pass {
-    group = "org.glassfish.jersey.core"
-    module = "jersey-server"
-    versions = "[2.0,]"
-  }
-}
-apply from: "${rootDir}/gradle/java.gradle"
-
-dependencies {
-  compileOnly group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.0'
-  compileOnly group: 'org.glassfish.jersey.core', name: 'jersey-server', version: '2.0'
-
-  compile project(':instrumentation:jax-rs-annotations-2.0')
-}
