--- instrumentation/jax-rs-client-1.1/src/main/java/io/opentelemetry/auto/instrumentation/jaxrs/v1/JaxRsClientV1Decorator.java
@@ -1,60 +1,0 @@
-/*
- * Copyright 2020, OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.opentelemetry.auto.instrumentation.jaxrs.v1;
-
-import com.sun.jersey.api.client.ClientRequest;
-import com.sun.jersey.api.client.ClientResponse;
-import io.opentelemetry.OpenTelemetry;
-import io.opentelemetry.auto.decorator.HttpClientDecorator;
-import io.opentelemetry.trace.Tracer;
-import java.net.URI;
-
-public class JaxRsClientV1Decorator extends HttpClientDecorator<ClientRequest, ClientResponse> {
-  public static final JaxRsClientV1Decorator DECORATE = new JaxRsClientV1Decorator();
-
-  public static final Tracer TRACER =
-      OpenTelemetry.getTracerFactory().get("io.opentelemetry.auto.jax-rs-client-1.1");
-
-  @Override
-  protected String getComponentName() {
-    return "jax-rs.client";
-  }
-
-  @Override
-  protected String method(final ClientRequest httpRequest) {
-    return httpRequest.getMethod();
-  }
-
-  @Override
-  protected URI url(final ClientRequest httpRequest) {
-    return httpRequest.getURI();
-  }
-
-  @Override
-  protected String hostname(final ClientRequest httpRequest) {
-    return httpRequest.getURI().getHost();
-  }
-
-  @Override
-  protected Integer port(final ClientRequest httpRequest) {
-    return httpRequest.getURI().getPort();
-  }
-
-  @Override
-  protected Integer status(final ClientResponse clientResponse) {
-    return clientResponse.getStatus();
-  }
-}
