--- instrumentation/jax-rs-client-2.0/connection-error-handling-jersey/connection-error-handling-jersey.gradle
@@ -1,15 +1,0 @@
-muzzle {
-  pass {
-    group = "org.glassfish.jersey.core"
-    module = "jersey-client"
-    versions = "[2.0,)"
-  }
-}
-
-apply from: "${rootDir}/gradle/java.gradle"
-
-dependencies {
-  compileOnly group: 'org.glassfish.jersey.core', name: 'jersey-client', version: '2.0'
-
-  compile project(':instrumentation:jax-rs-client-2.0')
-}
