--- instrumentation/jax-rs-client-2.0/connection-error-handling-resteasy/connection-error-handling-resteasy.gradle
@@ -1,15 +1,0 @@
-muzzle {
-  pass {
-    group = "org.jboss.resteasy"
-    module = "resteasy-client"
-    versions = "[2.0,)"
-  }
-}
-
-apply from: "${rootDir}/gradle/java.gradle"
-
-dependencies {
-  compileOnly group: 'org.jboss.resteasy', name: 'resteasy-client', version: '3.0.0.Final'
-
-  compile project(':instrumentation:jax-rs-client-2.0')
-}
