+++ instrumentation/jaxrs-1.0/src/test/groovy/JerseyTest.groovy
