+++ instrumentation/jaxrs-1.0/src/test/java/Resource.java
