+++ instrumentation/jaxrs-2.0/jaxrs-2.0-resteasy-3.0/jaxrs-2.0-resteasy-3.0.gradle
@@ -1,0 +1,25 @@
+muzzle {
+  // Cant assert fails because muzzle assumes all instrumentations will fail
+  // Instrumentations in jaxrs-2.0 will pass
+
+  // Resteasy changes a class's package in 3.1.0 then moves it back in 3.5.0 
+  pass {
+    group = "org.jboss.resteasy"
+    module = "resteasy-jaxrs"
+    versions = "[3.0.0.Final,3.1.0.Final)"
+  }
+
+  pass {
+    group = "org.jboss.resteasy"
+    module = "resteasy-jaxrs"
+    versions = "[3.5.0.Final,)"
+  }
+}
+apply from: "${rootDir}/gradle/java.gradle"
+
+dependencies {
+  compileOnly group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.0'
+  compileOnly group: 'org.jboss.resteasy', name: 'resteasy-jaxrs', version: '3.0.0.Final'
+
+  compile project(':instrumentation:jaxrs-2.0')
+}
