+++ instrumentation/jaxrs-2.0/src/test/groovy/JavaInterfaces.java
