+++ instrumentation/jaxrs-client-1.1/src/main/java/io/opentelemetry/auto/instrumentation/jaxrs/v1/InjectAdapter.java
