+++ instrumentation/jaxrs-client-1.1/src/test/groovy/JaxRsClientV1Test.groovy
