+++ instrumentation/jaxrs-client-2.0/jaxrs-client-2.0-jersey-2.0/jaxrs-client-2.0-jersey-2.0.gradle
@@ -1,0 +1,15 @@
+muzzle {
+  pass {
+    group = "org.glassfish.jersey.core"
+    module = "jersey-client"
+    versions = "[2.0,)"
+  }
+}
+
+apply from: "${rootDir}/gradle/java.gradle"
+
+dependencies {
+  compileOnly group: 'org.glassfish.jersey.core', name: 'jersey-client', version: '2.0'
+
+  compile project(':instrumentation:jaxrs-client-2.0')
+}
