+++ instrumentation/jaxrs-client-2.0/jaxrs-client-2.0-jersey-2.0/src/main/java/io/opentelemetry/auto/instrumentation/connection_error/jersey/JerseyClientConnectionErrorInstrumentation.java
