+++ instrumentation/jaxrs-client-2.0/jaxrs-client-2.0.gradle
@@ -1,0 +1,49 @@
+muzzle {
+  pass {
+    group = "javax.ws.rs"
+    module = "javax.ws.rs-api"
+    versions = "[2.0,)"
+  }
+  pass {
+    // We want to support the dropwizard clients too.
+    group = 'io.dropwizard'
+    module = 'dropwizard-client'
+    versions = "[0.8.0,)"
+    assertInverse = true
+  }
+}
+
+apply from: "${rootDir}/gradle/java.gradle"
+
+apply plugin: 'org.unbroken-dome.test-sets'
+
+testSets {
+  latestDepTest {
+    dirName = 'test'
+  }
+}
+
+dependencies {
+  compileOnly group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.0.1'
+  compileOnly group: 'javax.annotation', name: 'javax.annotation-api', version: '1.2'
+
+  testCompile project(':instrumentation:java-concurrent')
+
+  testCompile project(':instrumentation:jaxrs-client-2.0:jaxrs-client-2.0-jersey-2.0')
+  testCompile project(':instrumentation:jaxrs-client-2.0:jaxrs-client-2.0-resteasy-2.0')
+
+  testCompile group: 'javax.ws.rs', name: 'javax.ws.rs-api', version: '2.0.1'
+
+  testCompile group: 'org.glassfish.jersey.core', name: 'jersey-client', version: '2.0'
+  testCompile group: 'org.jboss.resteasy', name: 'resteasy-client', version: '3.0.0.Final'
+  testCompile group: 'org.apache.cxf', name: 'cxf-rt-rs-client', version: '3.1.0'
+  // Doesn't work with CXF 3.0.x because their context is wrong:
+  // https://github.com/apache/cxf/commit/335c7bad2436f08d6d54180212df5a52157c9f21
+
+  testCompile group: 'javax.xml.bind', name: 'jaxb-api', version: '2.2.3'
+
+  latestDepTestCompile group: 'org.glassfish.jersey.inject', name: 'jersey-hk2', version: '2.27'
+  latestDepTestCompile group: 'org.glassfish.jersey.core', name: 'jersey-client', version: '2.27'
+  latestDepTestCompile group: 'org.apache.cxf', name: 'cxf-rt-rs-client', version: '3.2.6'
+  latestDepTestCompile group: 'org.jboss.resteasy', name: 'resteasy-client', version: '3.0.26.Final'
+}
