+++ instrumentation/jaxrs-client-2.0/src/test/groovy/JaxMultithreadedClientTest.groovy
