+++ settings.gradle
@@ -77,15 +77,15 @@ include ':instrumentation:hibernate:core-4.0'
 include ':instrumentation:hibernate:core-4.3'
 include ':instrumentation:http-url-connection'
 include ':instrumentation:hystrix-1.4'
-include ':instrumentation:jax-rs-annotations-1.0'
-include ':instrumentation:jax-rs-annotations-2.0'
-include ':instrumentation:jax-rs-annotations-2.0:filter-jersey-2.0'
-include ':instrumentation:jax-rs-annotations-2.0:filter-resteasy-3.0'
-include ':instrumentation:jax-rs-annotations-2.0:filter-resteasy-3.1'
-include ':instrumentation:jax-rs-client-1.1'
-include ':instrumentation:jax-rs-client-2.0'
-include ':instrumentation:jax-rs-client-2.0:connection-error-handling-jersey'
-include ':instrumentation:jax-rs-client-2.0:connection-error-handling-resteasy'
+include ':instrumentation:jaxrs-1.0'
+include ':instrumentation:jaxrs-2.0'
+include ':instrumentation:jaxrs-2.0:jaxrs-2.0-jersey-2.0'
+include ':instrumentation:jaxrs-2.0:jaxrs-2.0-resteasy-3.0'
+include ':instrumentation:jaxrs-2.0:jaxrs-2.0-resteasy-3.1'
+include ':instrumentation:jaxrs-client-1.1'
+include ':instrumentation:jaxrs-client-2.0'
+include ':instrumentation:jaxrs-client-2.0:jaxrs-client-2.0-jersey-2.0'
+include ':instrumentation:jaxrs-client-2.0:jaxrs-client-2.0-resteasy-2.0'
 include ':instrumentation:java-concurrent'
 include ':instrumentation:java-concurrent:kotlin-testing'
 include ':instrumentation:java-concurrent:scala-testing'
